feat: install react-tagify lib, add config to axios requisition and move context.js to context folder
