+++ package.json
@@ -13,6 +13,7 @@
     "react-loader-spinner": "^5.3.4",
     "react-router-dom": "^6.6.1",
     "react-scripts": "5.0.1",
+    "react-tagify": "^0.1.4",
     "styled-components": "^5.3.6",
     "sweetalert": "^2.1.2",
     "web-vitals": "^2.1.4"
