--- src/contexts/modelo.js
@@ -1,1 +1,0 @@
-// CRIEI ESTE ARQUIVO SÓ PRA APARECER NO GITHUB :)
